+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Include process and thread IDs in log messages
 
 ### Dependencies
+- Use appropriate `requests-retry-session` version based on Python version
 - Fix `wheel` version to prevent build failures
 
 ## [1.27.1] - 09/06/2024